First step light client
