+++ package.json
@@ -26,8 +26,8 @@
     "build:electron": "yarn build:ui && lerna run build --scope=@substrate/electron-app",
     "build:extension": "INLINE_RUNTIME_CHUNK=false yarn build:ui && lerna run build --scope=@substrate/extension-app",
     "clean": "polkadot-dev-clean-build",
-    "fetch:substrate": "cd ./scripts && node fetchLatestSubstrate.js",
-    "lint": "tsc --noEmit && eslint --ext ts,tsx .",
+    "fetch:substrate": "cd ./packages/electron-app/scripts && node fetchLatestSubstrate.js",
+    "lint": "tsc --noEmit && eslint --ext js,ts,tsx .",
     "package:electron": "yarn build:electron && lerna run package --scope=@substrate/electron-app",
     "package:extension": "yarn build:extension && lerna run package --scope=@substrate/extension-app",
     "run:electron": "yarn build:electron && lerna run electron --scope=@substrate/electron-app",
