+++ packages/electron-app/src/main/app/substrateProcess.ts
@@ -8,15 +8,17 @@ import { bundledPath, logger } from '../util';
 
 let substrateProc: ChildProcess;
 
-// TEMPORARY: change to runSubstrateLight once the light client is available.
-export const runSubstrateDev = (): void => {
+/**
+ * Create a process and run the bundled substrate node
+ */
+export function runSubstrate(): void {
   if (substrateProc) {
     logger.error('Unable to initialise Parity Substrate more than once');
     return;
   }
 
-  logger.info('Running `substrate --dev`');
-  const substrate = spawn(bundledPath, ['--dev']); // FIXME: --light
+  logger.info('Running `substrate --light`');
+  const substrate = spawn(bundledPath, ['--light']);
 
   substrate.stdout.on('data', data => {
     logger.info(data.toString());
@@ -33,11 +35,11 @@ export const runSubstrateDev = (): void => {
   });
 
   substrateProc = substrate;
-};
+}
 
-export const killSubstrate = (): void => {
+export function killSubstrate(): void {
   if (substrateProc) {
     logger.info('Killing Substrate');
     substrateProc.kill();
   }
-};
+}
